+++ src/main/groovy/ru/vyarus/gradle/plugin/animalsniffer/report/ReportMessage.groovy
@@ -1,0 +1,17 @@
+package ru.vyarus.gradle.plugin.animalsniffer.report
+
+import groovy.transform.CompileStatic
+
+/**
+ * Parsed animalsniffer error.
+ *
+ * @author Vyacheslav Rusakov
+ * @since 19.03.2017
+ */
+@CompileStatic
+class ReportMessage {
+    String source
+    String line
+    String code
+    boolean parseFail
+}