+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 * Prevent potential intercepting output of other ant tasks (#3)
+* Change console reporting format (to match gradle-quality-plugin format):
+    - Each error split to source and code lines followed by empty line
+    - IDE will be able to show link for class reference (clickable) when line number available
 
 ### 1.2.0 (2016-08-27)
 * Gradle 3.0 compatibility